refactor(tests): split test_tools_load_recipe.py — extract TestMigrateRecipe to test_tools_migrate_recipe.py
Move TestMigrateRecipe (11 tests) to a new focused file, share _MINIMAL_SCRIPT_YAML
via tests/server/_helpers.py, and prune now-unused imports from test_tools_load_recipe.py.
Update tests/server/CLAUDE.md to reflect the new file.
